[package] support pre-md5summed passwords in 6in4 scripts
The he.net ipv6 tunnels require that a user supply the md5sum of his password
when updating the ipv4 tunnel endpoint. the 6in4 package supports this well,
but requires that a user store his password in plaintext in uci. this patch
tries to detect whether the password in uci is already an md5sum
(32 characters containing only [a-z0-9]), so users can avoid storing plaintext
passwords.

Signed-off-by: Steven Brudenell <steven.brudenell@gmail.com>

git-svn-id: svn://svn.openwrt.org/openwrt/trunk@22889 3c298f89-4303-0410-b956-a3cf2f4a3e73