switch: ROBO Switch Gigabit MII Support
I wrote this patch some time ago because I had a need for one of the
Gigabit ports (Linksys E3000) to be forced to 100FD.  This is based
on the robocfg sources included w/ the RT-N16 sources from ASUS.
Since work is progressing on a BGMAC driver that could be included in
OpenWRT, this may be useful to someone else.

In testing, forcing the speed to 10/100 or 1000 worked fine; however,
when trying to force full-duplex mode, the result was always
half-duplex.  I was not able to isolate the source of the problem
(this patch, driver or H/W limitation).  The only way I could get it
to work was to set the port to Auto, but then only advertise 100FD
(not included in this patch).

I have a modified version of the robocfg package as well, I'd have to
clean it up a little first (remove the full-duplex hack) before
submitting it if there is interest.

Signed-off-by: Nathan Hintz <nlhintz@hotmail.com>


git-svn-id: svn://svn.openwrt.org/openwrt/trunk@34992 3c298f89-4303-0410-b956-a3cf2f4a3e73