ramips: introduce DTS for ALL5003
Prior to the switch to 3.8, ALL5002 and ALL5003 boards shared mach-all5002.c.
With the change to DTS, this is no longer accurate, as ALL5002 is based on
Rt3352 SoC while ALL5003 is based on Rt5350 SoC, therefore the corresponding
rt5350.dtsi need to be included for ALL5003.

Signed-off-by: Daniel Golle <dgolle@allnet.de>
Patchwork: http://patchwork.openwrt.org/patch/3557/
Acked-by: John Crispin <blogic@openwrt.org>

git-svn-id: svn://svn.openwrt.org/openwrt/trunk@36435 3c298f89-4303-0410-b956-a3cf2f4a3e73