6rd: Make don't fragment bit (df) a configurable variable
Make the IPv4 "do not fragment" bit a configurable variable for the 6rd
tunnel interface as the bit should not be set in the encapsulating IPv4
header according to RFC3056.
On top the config variable allows to enable/disable path mtu discovery
for tunnel interfaces.
Config variable can be passed to netifd as tunnel config support is
already present for the "do not fragment" bit.

Signed-off-by: Hans Dedecker <dedeckeh@gmail.com>

git-svn-id: svn://svn.openwrt.org/openwrt/trunk@39370 3c298f89-4303-0410-b956-a3cf2f4a3e73