ath: Add config option for DFS support
This patch adds a config option to enable DFS in ath9k and ath10k
drivers (with ath10k you need the AP-branch firmware). I'm not entirely
sure about the regulatory compliance issues, though. As far as I
understand, enabling the DFS_CERTIFIED config options implies that the
device is certified for DFS. No doubt the original firmware has obtained
such a certification, but not with the open source drivers. I suppose
that as long as this is disabled in default builds (like ATH_USER_REGD)
everything should be fine, but it would be nice to have out-of-the-box
OpenWrt support for DFS. I'm also not sure whether the description of
the configuration option should say something about regulatory compliance.

Signed-off-by: Matti Laakso <malaakso@elisanet.fi>
Signed-off-by: Felix Fietkau <nbd@openwrt.org>

git-svn-id: svn://svn.openwrt.org/openwrt/trunk@40692 3c298f89-4303-0410-b956-a3cf2f4a3e73