ar71xx: Add support for MikroTik RouterBOARD SXT Lite
This patch adds support for the MikroTik RouterBOARD SXT Lite.

The new RB911L series is also supported as a result.

v2 of this patch fixes the wmac offset to match what is on the sticker.
v3 refreshes the patch against r41148 and defines the power led as the status led in diag.sh
v4 refreshes the patch against r41353 and fixes the patch path issue to make git am work correctly
v5 selects the kernel config option in the mikrotik profile rather than in the main ar71xx config

Signed-off-by: Matthew Reeve <mreeve@tenxnetworks.com>
Signed-off-by: John Crispin <blogic@openwrt.org>

git-svn-id: svn://svn.openwrt.org/openwrt/trunk@41450 3c298f89-4303-0410-b956-a3cf2f4a3e73