toolchain: musl: fix symlink for ldd
ld-musl-*.so* is a symlink "broken" for the hostsystem, so wildcard
will skip it, causing LD_MUSL_NAME to empty and the ldd symlink pointing
to ../../lib directly.
This causes sysupgrade failing to copy any linked libaries and
consequently failing to run anything after switching to ram disk.

Fix this by creating a symlink directly pointing to where ld-musl-*.so*
points to.

Signed-off-by: Jonas Gorski <jogo@openwrt.org>

git-svn-id: svn://svn.openwrt.org/openwrt/trunk@43314 3c298f89-4303-0410-b956-a3cf2f4a3e73