version.mk: add "%s" placeholder for dotted target.subtarget notation
Implement "%s" placeholder that expands to either the target name,
e.g. "ar71xx" if the subtarget is generic or to target.subtarget, e.g.
"ar71xx.nand" is a subtarget is choosen.

Also change the default repository url template to use "%s" instead
of "%T" to reflect the directory structure used by the buildbot systems.

Signed-off-by: Jo-Philipp Wich <jow@openwrt.org>

git-svn-id: svn://svn.openwrt.org/openwrt/trunk@43871 3c298f89-4303-0410-b956-a3cf2f4a3e73