ipq806x: ap148/r7500: fix eth0 for non gige speeds
Eth0 is attached to mac0 of the switch with a fixed link and and not to
phy4 in single phy mode, so configuring it to anything but 1000FD will
break the connection, which will happen if a only 100 Mbit capapble device
is plugged into the wan port.

Fix this by not taking the state from phy4 and just configuring a fixed link
for eth0.

Signed-off-by: Jonas Gorski <jogo@openwrt.org>

git-svn-id: svn://svn.openwrt.org/openwrt/trunk@47695 3c298f89-4303-0410-b956-a3cf2f4a3e73