kernel: drop old uimage splitter from 4.4
For a long time we have a dedicated firmware splitter for uimage as a
replacement for the one builtin into mtd subsystem. For a year we got a
proper warning for users sticking to the old one (r43700). It should be
really safe to drop it now, especially since we do it for 4.4 only.

Signed-off-by: Rafał Miłecki <zajec5@gmail.com>

git-svn-id: svn://svn.openwrt.org/openwrt/trunk@47747 3c298f89-4303-0410-b956-a3cf2f4a3e73