ramips: add additional support for MQmaker WiTi board
The board is based on mt7621AT cpu, and has 16mb nor flash, 256mb of ram,
2 sata ports, microsd card slot, 1 USB 3.0 port and at least one 2.4 and
one 5 ghz antenna.

This is the 6th submission that adds support for XHCI in the device tree
file, along with switching the location of the 2 radio's and addition of
the kmod-i2c-mt7621 in the default packages of the profile.

Signed-off-by: Sebastian Careba <nitroshift@yahoo.com>

git-svn-id: svn://svn.openwrt.org/openwrt/trunk@47845 3c298f89-4303-0410-b956-a3cf2f4a3e73