ramips: Add support for Phicomm PSG1208
This patch adds support for Phicomm PSG1208.This is a router with MT7620A SoC with 8M flash and 64M ram.
The WPS led is uesd as status_led because the power light can't be controlled with GPIO.
It seems that the 5g wifi led is connected to MT7612E and it can't be controlled with GPIO too.

Signed-off-by: Chuanhong Guo <gch981213@gmail.com>

git-svn-id: svn://svn.openwrt.org/openwrt/trunk@48721 3c298f89-4303-0410-b956-a3cf2f4a3e73